Fixed parsing bug, began adding 'fromScript' support, and added test.
Fixed a bug in the AFTER fragment that used a token's 'parse' method instead of the parseContext's 'parse'. Also added the 'fromScript' method to a few fragments to allow conversions from ASM to miniscript strings. Jest was added as a test runner and a single 'fromScript' test has been added.